Migrate CustomFileInput to TypeScript
